chore(ssi): move functions to the files they are called in
This commit moves functions out of the `auto_instrumentation` file to
where they are called. This file had grown organically and is no longer
relevant.
