+++ pkg/clusteragent/admission/mutate/autoinstrumentation/target_mutator.go
@@ -484,3 +484,14 @@ func getEnabledLabel(pod *corev1.Pod) (bool, bool) {
 
 	return false, found
 }
+
+// getAllLatestDefaultLibraries returns all supported by APM Instrumentation tracing libraries
+// that should be enabled by default
+func getAllLatestDefaultLibraries(containerRegistry string) []libInfo {
+	var libsToInject []libInfo
+	for _, lang := range supportedLanguages {
+		libsToInject = append(libsToInject, lang.defaultLibInfo(containerRegistry, ""))
+	}
+
+	return libsToInject
+}
